the feature 'Asset Locations' learn to manage multiple locations
Before in dandelion.properties, 'assetsLocations=remote' only retrieve the location 'remote' for an asset.
Now, 'assetsLocations=remote,local' retrieve the location 'remote' also but if this location is undefined, it's retrieve location 'local'.

In case of Override Asset in a new scope, the locations are merged between scopes when you want access them.
Now, the case of add a new location for an asset without change the other locations, it's easy.