[Core][EhCache] Added distinction between cache and storage
Prior to this commit, the implementation of AssetCache was used to cache both asset contents and the association between a request and its corresponding assets (metadata only). Without taking take, a sync had been introduced between these 2 caches, thus leading to NPE in some cases.
Asset contents are now stored in a AssetStorage (a new MemoryAssetStorage has been added, configured by default) which is not considered as a cache any longer, but as a data storage needed to render the assets served by the DandelionServlet. A custom implementation can be provided thanks to the corresponding SPI.
Many changes have been done in lots of classes accordingly.

* Filtering of assets made in the AssetQuery has been improved
* (.*)AssetCache renamed to (.*)Cache for more consistency with the new AssetStorage
* Added ASSET_STORAGE as a new configuration point allowing to select the implementation of AssetStorage
* Renamed ASSET_CACHING configuration point to CACHE
* Removed CACHE_ASSET_MAX_SIZE
* Renamed CACHE_REQUEST_MAX_SIZE to CACHE_MAX_SIZE
* EhCache extra updated to comply with the new caching API
* Updated all tests
* Added comments

Fixes #47